Add files via upload
In this project, my aim is to find the most efficient model describing the Mean Temperature of the day, given the past three days of weather measurements, current month, current day by using Madrid_Weather_1991_1995.csv. Since redundancy and missing occurrences in the data, I spent most of my time on the preprocessing of the data. At the latest stage of data-preprocessing, coming up with the most meaningful and best explaining features was aimed. Later on, by trying different sets of models in regression, I fitted the most efficient 3 regressor models and optimized them.